pyproject.toml
[project]
name = "grpcmcp"
version = "0.1.0"
description = "gRPC transport for Model Context Protocol"
license = { text = "MIT" }
dependencies = [
    "grpcio>=1.78.0",
    "protobuf>=5.26.1",
    "mcp",
    "mcp-transport-proto",
    "types-protobuf>=6.32.1.20251210",
    "pydantic>=2.12.5",
    "grpcio-reflection>=1.78.0",
]
requires-python = ">=3.10"


[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.pyright]
venvPath = "."
venv = ".venv"
include = ["src", "tests", "example"]

[tool.uv.sources]
mcp = { git = "https://github.com/pawbhard/python-sdk.git", branch = "extract-dispatcher-from-base-session" }
mcp-transport-proto = { git = "https://github.com/GoogleCloudPlatform/mcp-grpc-transport-proto.git" }

[dependency-groups]
dev = [
    "pyright>=1.1.408",
    "ruff>=0.14.13",
    "pytest>=8.0.0",
    "pylint>=4.0.5",
]
